# server/docs/openapi.build.yaml
openapi: 3.0.0
info:
  title: StartKit.AI API
  version: 1.0.0
  description: >
    # Introduction


    The StartKit.AI API allows you to create a starting point for your AI project

    very quickly.


    It's emphasis is on simplicity, so we've tried to make the API as easy as possible

    to get started with.


    All OpenAI API endpoints that result in a query to their LLM models are supported.

    But the APIs that do account level things such as creating assistants or fine-tunes

    are not implemented, as we either consider them not useful, or replaced by other

    StartKit.AI functionality.



    ## Endpoints


    Our AI endpoints range from simple to complex wrappers around OpenAI's API

    functionality.


    They give you all the things that are missing when trying to build a SaaS AI app.


    These API pages give you details of the arguments and details needed to call

    the API. Descriptions of how the endpoints work and their justifications can be

    found with the [main website documentation here](https://startkit.ai/docs).


    ### [Chat](#tag/Chat)


    The main functionality for most AI applications, our Chat endpoints gives you total

    control when allowing your customers to comminicate with a LLM.


    - Saves chat history to the database to build longer and smarter context windows

    - Hooks into pre-stored embeddings via context IDs.

    - Uploads and stores created images

    - Fetches files and webpages automatically


...


...


...


...

...
servers:
...
paths:
...
  /api/moderation/harmful:
    post:
      summary: Check if text is harmful
      operationId: checkHarmfulText
      description: >-
        Check if some given text is considered harmful. One or more harmful
        categories will be returned.


        ### Usage


        The harmful text API doesn’t count towards to cost usage limits, as OpenAI actually offers this API for free.
      tags:
        - Moderation
      security:
        - bearer_auth: []
      parameters:
        - $ref: "#/paths/~1api~1chat/post/parameters/0"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
              required:
                - text
            example:
              text: I want to kill them
      responses:
        "200":
          description: >-
            Harmfulness analysis result. The result contains a boolean
            `flagged`, and all

            the categories that are flagged from the following list.

            ```json

            [
              "sexual",
              "hate",
              "harassment",
              "self-harm",
              "sexual/minors",
              "hate/threatening",
              "violence/graphic",
              "self-harm/intent",
              "self-harm/instructions",
              "harassment/threatening",
              "violence"
            ]

            ```
          content:
            application/json:
              schema:
                type: object
                properties:
                  flagged:
                    type: boolean
                    example: true
                  categories:
                    type: array
                    items:
                      enum:
                        - violence
                        - sexual
                        - hate
                        - harassment
                        - self-harm
                        - sexual/minors
                        - hate/threatening
                        - violence/graphic
                        - self-harm/intent
                        - self-harm/instructions
                        - harassment/threatening
        "400":
          description: Bad request, text is required
        "401":
          description: Unauthorized, user not authenticated.
        "500":
          description: An error occurred.
  /api/moderation/harmful/categories:
    get:
      summary: Retrieve harmful categories
      operationId: getHarmfulCategories
      tags:
        - Moderation
      security:
        - bearer_auth: []
      parameters:
        - $ref: "#/paths/~1api~1chat/post/parameters/0"
      description: |-
        Retrieve the categories used for harmful text.
        ```js
        [
          "sexual",
          "hate",
          "harassment",
          "self-harm",
          "sexual/minors",
          "hate/threatening",
          "violence/graphic",
          "self-harm/intent",
          "self-harm/instructions",
          "harassment/threatening",
          "violence"
        ]
        ```
      responses:
        "200":
          description: Fetch a list of harmful content categories.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                  - sexual
                  - hate
                  - harassment
                  - self-harm
                  - sexual/minors
                  - hate/threatening
                  - violence/graphic
                  - self-harm/intent
                  - self-harm/instructions
                  - harassment/threatening
                  - violence
        "401":
          description: Unauthorized, user not authenticated.
        "500":
          description: An error occurred.
  /api/moderation/sentiment:
    post:
      summary: Get sentiment of text
      operationId: getSentiment
      tags:
        - Moderation
      description: Get sentiment of text
      security:
        - bearer_auth: []
      parameters:
        - $ref: "#/paths/~1api~1chat/post/parameters/0"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
              required:
                - text
            example:
              text: Ugh, this new chef sucks. I'll never eat here again
      responses:
        "200":
          description: Sentiment analysis result with usage information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sentiment:
                    type: object
                    properties:
                      sentiment:
                        type: string
                        enum:
                          - positive
                          - negative
                          - neutral
                      confidence:
                        type: integer
                        format: int32
                        minimum: 0
                        maximum: 100
                    example:
                      sentiment: negative
                      confidence: 95
                  usage:
                    type: array
                    items:
                      type: object
                      properties:
                        model:
                          type: string
                        promptTokens:
                          type: integer
                          format: int32
                        completionTokens:
                          type: integer
                          format: int32
                        cost:
                          type: number
                          format: float
                    example:
                      model: gpt-4-turbo
                      promptTokens: 44
                      completionTokens: 989
                      cost: 0.012
        "400":
          description: Bad request, text is required
        "401":
          description: Unauthorized, user not authenticated.
        "500":
          description: An error occurred.
  /api/moderation/redact:
    post:
      summary: Redact sensitive information
      operationId: redactSensitiveText
      tags:
        - Moderation
      security:
        - bearer_auth: []
      parameters:
        - $ref: "#/paths/~1api~1chat/post/parameters/0"
      description: Redact sensitive information from given text.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
              required:
                - text
            example:
              text: Hello everyone! Anyone here familiar with the new Node.js update? Could
                you email me at james(at)squarecat.io if you know anything? Or
                you can phone me 082-34234-44
      responses:
        "200":
          description: Redaction result with usage information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      redactions:
                        type: array
                        items:
                          type: object
                          properties:
                            type:
                              type: string
                              enum:
                                - email_address
                                - phone_number
                                - name
                                - keyword
                                - custom
                            text:
                              type: string
                            confidence:
                              type: integer
                              format: int32
                              minimum: 0
                              maximum: 100
                              description: Confidence score of the redaction as a percentage.
                      redactedText:
                        type: string
                        description: Text after applying redactions with placeholders for redacted
                          content.
                    example:
                      redactedText: Hello everyone! Anyone here familiar with the new Node.js update?
                        Could you email me at [email_redacted] if you know
                        anything? Or you can phone me at [phone_redacted]
                      redactions:
                        - type: email_address
                          text: james(at)squarecat.io
                          confidence: 95
                        - type: phone_number
                          text: 082-34234-44
                          confidence: 90
                  usage:
                    $ref: "#/paths/~1api~1moderation~1sentiment/post/responses/200/content/applicat\
                      ion~1json/schema/properties/usage"
        "400":
          description: Bad request, text is required
        "401":
          description: Unauthorized, user not authenticated.
        "500":
          description: An error occurred.
...
  /api/speech/from-text:
    post:
      summary: Create speech from text
      operationId: createSpeechFromText
      tags:
        - Speech
      security:
        - bearer_auth: []
      parameters:
        - $ref: "#/paths/~1api~1chat/post/parameters/0"
      description: Create speech from text
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
            example:
              text: Seize the time... Live now! Make now always the most precious time. Now
                will never come again. -  Captain Jean-Luc Picard
      responses:
        "200":
          description: Returns an audio file
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/opus:
              schema:
                type: string
                format: binary
            audio/aac:
              schema:
                type: string
                format: binary
            audio/flac:
              schema:
                type: string
                format: binary
            audio/wav:
              schema:
                type: string
                format: binary
            audio/pcm:
              schema:
                type: string
                format: binary
        "400":
          description: Bad request, text is required
        "401":
          description: Unauthorized, user not authenticated.
        "500":
          description: An error occured
  /api/speech/translate/from-audio:
    post:
      summary: Translate audio to English
      operationId: translateAudio
      tags:
        - Speech
      description: Accepts an audio file and translates its content to the audio in
        the specified language.
      security:
        - bearer_auth: []
      parameters:
        - $ref: "#/paths/~1api~1chat/post/parameters/0"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - audio
              properties:
                audio:
                  type: string
                  format: binary
                  description: The audio file to be translated.
                prompt:
                  type: string
                  description: Additional prompt to guide the translation.
                  example: Translate this conversation to simple English from French
                language:
                  type: string
                  description: Language code for the language of the audio file.
                  example: fr
      responses:
        "200":
          description: A translated audio file.
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/opus:
              schema:
                type: string
                format: binary
            audio/aac:
              schema:
                type: string
                format: binary
            audio/flac:
              schema:
                type: string
                format: binary
            audio/wav:
              schema:
                type: string
                format: binary
            audio/pcm:
              schema:
                type: string
                format: binary
        "400":
          description: Bad request, audio file is required
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: File named 'audio' is required
        "401":
          description: Unauthorized, user not authenticated.
        "500":
          description: An error occured
  /api/speech/transcribe:
    post:
      summary: Transcribe audio content
      operationId: transcribeAudio
      description: Accepts an audio file, transcribes its content, and returns the
        transcription along with usage data.
      security:
        - bearer_auth: []
      parameters:
        - $ref: "#/paths/~1api~1chat/post/parameters/0"
      tags:
        - Speech
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - audio
              properties:
                audio:
                  type: string
                  format: binary
                  description: The audio file to be transcribed.
                language:
                  type: string
                  description: Optional language code of the audio content.
                  example: fr
                prompt:
                  type: string
                  description: Optional prompt to guide the transcription.
      responses:
        "200":
          description: Successfully transcribed audio content.
          content:
            application/json:
              schema:
                type: object
                properties:
                  transcription:
                    type: object
                    properties:
                      text:
                        type: string
                        description: The textual transcription of the audio file.
                        example: Hello, this is an example of transcribed audio.
                      segments:
                        type: array
                        example: []
                  usage:
                    type: object
                    description: Usage data related to the transcription request.
                    properties:
                      model:
                        type: string
                        description: The model used for the transcription
                        example: whisper-1
                      audioFileDurationSeconds:
                        type: integer
                        description: The duration of the transcribed audio
                        example: 18
                      cost:
                        type: number
                        description: Cost of the transcription
                        example: 0.05
        "400":
          description: Bad request, audio file is required
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: File named 'audio' is required
        "401":
          description: Unauthorized, user not authenticated.
        "500":
          description: An error occured
...
components:
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      description: This request requires a license key to be included in the
        Authorization header of the request.